refact : foodItem 의 price를 Int데이터 타입으로 변경
-데이터 타입 변경에 따라 BindingAdapters의 bigDecimalToPriceString의 함수 이름 수정
-item_food.xml의 함수이름 수정

+++ android/app/src/main/java/com/ssafy/foodfind/data/entity/FoodItem.kt
@@ -7,7 +7,7 @@ data class FoodItem (
 	val truckId : Int = 0,
 	var name : String = "",
 	var description : String = "",
-	var price : BigDecimal = BigDecimal(0),
+	var price : Int = 0,
 	var status : String = "",
 	val usingYN : Boolean = true
 ) : java.io.Serializable
